Move getDirPath to new file so that it can be re-exported for testing.
